requirements.txt
flask==2.0.1
requests==2.28.1
python-dotenv==0.19.2
